refactor: organizar archivos
